reorganize component imports for consistency
